Update GLFW, glm, spdlog and ImGui to HEAD
Visual Studio project for spdlo was created
GLFW, ImGui and glm is now a submodule

+++ Engine/src/External/Render/Vulkan/utils/Debug.cpp
@@ -177,7 +177,7 @@ namespace RT::Vulkan
 
 	void checkVkResultCallback(const VkResult result)
 	{
-		RT_ASSERT(VK_SUCCESS == result, "[Vulkan] Required immediate abort VkResult = {}", result);
+		RT_ASSERT(VK_SUCCESS == result, "[Vulkan] Required immediate abort VkResult = {}", (uint32_t)result);
 	}
 
 	std::vector<const char*> getRequiredExtensions()
