+++ Engine/src/External/Render/Vulkan/utils/Debug.h
@@ -56,7 +56,7 @@ namespace RT::Vulkan
     #define BEGIN_VK_DEBUG_LABEL(CMD_BUFF, LABEL_NAME, COLOR)                RT::Vulkan::beginDebugLabel(CMD_BUFF, LABEL_NAME, COLOR)
     #define END_VK_DEBUG_LABEL(CMD_BUFF)                                     RT::Vulkan::endDebugLabel(CMD_BUFF)
 
-    #define CHECK_VK(EXPR, MSG) { VkResult result = EXPR; if (VK_SUCCESS != result) { RT_LOG_ERROR("Result id = {}: " MSG, result); DEBUGBREAK; } }
+    #define CHECK_VK(EXPR, MSG) { VkResult result = EXPR; if (VK_SUCCESS != result) { RT_LOG_ERROR("Result id = {}: " MSG, (uint32_t)result); DEBUGBREAK; } }
 #else
     #define SET_VK_DEBUG_NAME()
     #define SET_VK_DEBUG_TAG()
